chore: Update HeroSection.js, Footer.js, and tailwind.config.js
This commit updates the HeroSection.js, Footer.js, and tailwind.config.js files. In HeroSection.js, the text size and spacing were adjusted for better visual alignment. In Footer.js, the text size was modified for improved readability. In tailwind.config.js, the font family was updated from Nunito_Sans to Nunito. These changes were made to enhance the design and functionality of the website.